backend/todolist/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
pytest==7.3.1
pytest-mock==3.10.0
pytest-print==0.3.1
pytz==2023.3.post1
sqlparse==0.4.4
